new tensor selection for all grids
