+++ SparseGrids/tsgIndexManipulator.hpp
@@ -71,30 +71,6 @@ namespace MultiIndexManipulations{
 /*!
  * \internal
  * \ingroup TasmanianMultiIndexManipulations
- * \brief Create a full-tensor multi-index set with \b num_entries in each direction.
- * \endinternal
- */
-inline MultiIndexSet generateFullTensorSet(std::vector<int> const &num_entries){
-    size_t num_dimensions = num_entries.size();
-    int num_total = 1;
-    for(auto &l : num_entries) num_total *= l;
-    std::vector<int> indexes(Utils::size_mult(num_dimensions, num_total));
-    auto iter = indexes.rbegin();
-    for(int i=num_total-1; i>=0; i--){
-        int t = i;
-        auto l = num_entries.rbegin();
-        // in order to generate indexes in the correct order, the for loop must go backwards
-        for(size_t j = 0; j<num_dimensions; j++){
-            *iter++ = (t % *l);
-            t /= *l++;
-        }
-    }
-    return MultiIndexSet(num_dimensions, indexes);
-}
-
-/*!
- * \internal
- * \ingroup TasmanianMultiIndexManipulations
  * \brief Generate the multi-index with entries satisfying the \b inside(), assumes that \b inside() defines a lower-complete set.
  * \endinternal
  */
@@ -120,103 +96,10 @@ inline MultiIndexSet generateLowerMultiIndexSet(size_t num_dimensions, std::func
 /*!
  * \internal
  * \ingroup TasmanianMultiIndexManipulations
- * \brief Generate a series of \b level_sets where each set has the parents/children of the previous one that satisfy the \b inside() criteria.
- *
- * On entry, \b level_sets must constain at least one set.
- * The function takes the last set in \b level_sets and adds a new set of
- * eithe the parents or children that also satisfy the \b inside() condition.
- * The process is repeated until the new set is empty.
- * \endinternal
- */
-template<bool use_parents>
-void repeatAddIndexes(std::function<bool(const std::vector<int> &index)> inside, std::vector<MultiIndexSet> &level_sets){
-    size_t num_dimensions = level_sets.back().getNumDimensions();
-    bool adding = true;
-    while(adding){
-        Data2D<int> level((int) num_dimensions, 0);
-        int num_indexes = level_sets.back().getNumIndexes();
-        for(int i=0; i<num_indexes; i++){
-            std::vector<int> point(num_dimensions);
-            std::copy_n(level_sets.back().getIndex(i), num_dimensions, point.data());
-            for(auto &p : point){
-                p += (use_parents) ? -1 : 1; // parents have lower index, children have higher indexes
-                if ( (!use_parents || (p >= 0)) && inside(point) ) level.appendStrip(point);
-                p -= (use_parents) ? -1 : 1; // restore p
-            }
-        }
-
-        adding = (level.getNumStrips() > 0);
-        if (adding)
-            level_sets.push_back(MultiIndexSet(level));
-    }
-}
-
-/*!
- * \internal
- * \ingroup TasmanianMultiIndexManipulations
- * \brief Retuns the union of all \b level_sets, all sets are destroyed in the process.
- * \endinternal
- */
-inline MultiIndexSet unionSets(std::vector<MultiIndexSet> &level_sets){
-    size_t num_levels = level_sets.size();
-    while(num_levels > 1){
-        size_t stride = num_levels / 2 + (((num_levels % 2) > 0) ? 1 : 0);
-        for(size_t i=0; i<stride; i++)
-            if (i + stride < num_levels)
-                level_sets[i].addMultiIndexSet(level_sets[i + stride]);
-        num_levels = stride;
-    }
-    return std::move(level_sets[0]);
-}
-
-/*!
- * \internal
- * \ingroup TasmanianMultiIndexManipulations
  * \brief Expand the \b set with the minimum number of multi-indexes that will result in a lower complete set.
  * \endinternal
  */
-inline void completeSetToLower(MultiIndexSet &set){
-    size_t num_dimensions = set.getNumDimensions();
-    int num = set.getNumIndexes();
-    Data2D<int> completion((int) num_dimensions, 0);
-    for(int i=0; i<num; i++){
-        std::vector<int> point(num_dimensions);
-        std::copy_n(set.getIndex(i), num_dimensions, point.data());
-        for(auto &p : point){
-            if (p != 0){
-                p--;
-                if (set.missing(point)) completion.appendStrip(point);
-                p++;
-            }
-        }
-    }
-
-    if (completion.getNumStrips() > 0){
-        std::vector<MultiIndexSet> level_sets = { MultiIndexSet(completion) };
-
-        repeatAddIndexes<true>([&](std::vector<int> const &p) -> bool{ return set.missing(p); }, level_sets);
-
-        set.addMultiIndexSet(unionSets(level_sets));
-    }
-}
-
-/*!
- * \internal
- * \ingroup TasmanianMultiIndexManipulations
- * \brief Generate the minimum lower complete multi-index set that includes the indexes satisfying \b criteria(), assumes \b criteria() defines a connected set.
- * \endinternal
- */
-inline MultiIndexSet generateGeneralMultiIndexSet(size_t num_dimensions, std::function<bool(const std::vector<int> &index)> criteria){
-    std::vector<int> root(num_dimensions, 0);
-    std::vector<MultiIndexSet> level_sets = { MultiIndexSet(num_dimensions, root) };
-
-    repeatAddIndexes<false>(criteria, level_sets);
-
-    MultiIndexSet set = unionSets(level_sets);
-
-    completeSetToLower(set);
-    return set;
-}
+void completeSetToLower(MultiIndexSet &set);
 
 /*!
  * \internal
@@ -278,135 +161,6 @@ struct ProperWeights{
 
 /*!
  * \internal
- * \brief Generate weights cache for the given parameters.
- *
- * Complexity chosen here in favor of performance and re-usability.
- * - \b isotropic indicates whether to treat the \b offset as a maximum cache size in each direction (\b true)
- *   of as a cut-off for the weights, i.e., cache until the 1-D weight exceeds the offset.
- * - \b weights.contour must match \b contour
- * - if \b contour is \b type_level, then \b CacheType must be \b int
- * - if \b contour is \b type_curved or \b type_hyperbolic, then \b CacheType must be \b double
- * \endinternal
- */
-template<typename CacheType, TypeDepth contour, bool isotropic>
-std::vector<std::vector<CacheType>> generateLevelWeightsCache(ProperWeights const &weights, std::function<int(int i)> rule_exactness, int offset){
-    size_t num_dimensions = weights.getNumDimensions();
-    std::vector<std::vector<CacheType>> cache(num_dimensions);
-    std::vector<int> exactness_cache;
-
-    // if we know the sized here, then assign, otherwise dynamically build later
-    if (isotropic){
-        for(auto &vec : cache) vec.reserve((size_t) offset);
-        exactness_cache.resize((size_t) offset);
-        exactness_cache[0] = 0;
-        for(int i=1; i<offset; i++)
-            exactness_cache[i] = 1 + rule_exactness(i - 1);
-    }else{
-        exactness_cache.push_back(0); // exactness always starts from 0
-    }
-
-    for(size_t j=0; j<num_dimensions; j++){
-        int wl = weights.linear[j]; // linear weights
-        double wc = (contour == type_level) ? 0.0 : weights.curved[j]; // curved weights
-
-        size_t i = 0; // keep track of the index
-        CacheType w = (CacheType) (contour == type_hyperbolic) ? 1 : 0;
-        cache[j].push_back(w); // initial entry
-        do{ // accumulate the cache
-            i++;
-            if (!isotropic && (i >= exactness_cache.size()))
-                exactness_cache.push_back(1 + rule_exactness((int)(i - 1)));
-
-            int e = exactness_cache[i];
-
-            if (contour == type_level){
-                w = wl * e;
-            }else if (contour == type_curved){
-                w = (CacheType)(wl * e) + wc * log1p((CacheType) e);
-            }else{ // must be hyperbolic
-                w = pow((CacheType) (1 + e), wc);
-            }
-
-            cache[j].push_back(w);
-        // isotropic mode works until offset, anisotropic mode works until weight reaches the offset
-        }while( (!isotropic && (std::ceil(w) <= (CacheType) offset)) || (isotropic && (i < (size_t) offset)) );
-    }
-
-    return cache;
-}
-
-/*!
- * \internal
- * \brief Generate the multi-index of indexes with weighs less than the \b normalized_offset.
- *
- * The weight of an index uses the \b weights combined with the \b rule_exactness().
- * Called only when the set is guaranteed to be lower complete,
- * then the one dimensional weights can be cached prior to running the selection algorithm.
- * \endinternal
- */
-inline MultiIndexSet selectLowerSet(ProperWeights const &weights, std::function<int(int i)> rule_exactness, int normalized_offset){
-    size_t num_dimensions = weights.getNumDimensions();
-    if (weights.contour == type_level){
-        auto cache = generateLevelWeightsCache<int, type_level, false>(weights, rule_exactness, normalized_offset);
-        return generateLowerMultiIndexSet(num_dimensions,
-                [&](std::vector<int> const &index)->bool{
-                    int w = 0;
-                    for(size_t j=0; j<num_dimensions; j++) w += cache[j][index[j]];
-                    return (w <= normalized_offset);
-                });
-    }else if (weights.contour == type_curved){
-        auto cache = generateLevelWeightsCache<double, type_curved, false>(weights, rule_exactness, normalized_offset);
-        double noff = (double) normalized_offset;
-        return generateLowerMultiIndexSet(num_dimensions,
-                [&](std::vector<int> const &index)->bool{
-                    double w = 0;
-                    for(size_t j=0; j<num_dimensions; j++) w += cache[j][index[j]];
-                    return (std::ceil(w) <= noff);
-                });
-    }else{ // type_hyperbolic
-        auto cache = generateLevelWeightsCache<double, type_hyperbolic, false>(weights, rule_exactness, normalized_offset);
-        double noff = (double) normalized_offset;
-        return generateLowerMultiIndexSet(num_dimensions,
-                [&](std::vector<int> const &index)->bool{
-                    double w = 1.0;
-                    for(size_t j=0; j<num_dimensions; j++) w *= cache[j][index[j]];
-                    return (std::ceil(w) <= noff);
-                });
-    }
-}
-
-/*!
- * \internal
- * \ingroup TasmanianMultiIndexManipulations
- * \brief Generates the minimum lower complete set that contains all indexes with weights less than \b normalized_offset.
- *
- * The weight of an index uses the \b weights combined with the \b rule_exactness().
- * Called only for contour \b type_curved and caches values on-the-fly.
- * \endinternal
- */
-inline MultiIndexSet selectGeneralSet(ProperWeights const &weights, std::function<int(int i)> rule_exactness, int normalized_offset){
-    size_t num_dimensions = weights.getNumDimensions();
-    std::vector<std::vector<double>> cache(num_dimensions);
-    for(size_t j=0; j<num_dimensions; j++) cache[j].push_back(0.0);
-    double noff = (double) normalized_offset;
-    return generateGeneralMultiIndexSet(num_dimensions,
-                                        [&](std::vector<int> const &index) -> bool{
-                                            double w = 0;
-                                            for(size_t j=0; j<num_dimensions; j++){
-                                                while(index[j] >= (int) cache[j].size()){
-                                                    int exactness = 1 + rule_exactness((int)(cache[j].size() - 1));
-                                                    cache[j].push_back( (double) weights.linear[j] * exactness +
-                                                                         weights.curved[j] * log1p((double) exactness) );
-                                                }
-                                                w += cache[j][index[j]];
-                                            }
-                                            return (std::ceil(w) <= noff);
-                                        });
-}
-
-
-/*!
- * \internal
  * \brief Generate a lower complete multi-index set that satisfies the given properties.
  *
  * Using the \b num_dimensions select all multi-indexes with weight less than the \b offset.
@@ -414,49 +168,8 @@ inline MultiIndexSet selectGeneralSet(ProperWeights const &weights, std::functio
  * Rule exactness is used in place of the raw-index to cover the correct polynomial space, see references in \b TypeDepth.
  * \endinternal
  */
-MultiIndexSet selectTensors(size_t num_dimensions, int offset, TypeDepth type, std::function<int(int i)> rule_exactness, std::vector<int> const &anisotropic_weights);
-
-//! \internal
-//! \brief Set an empty vector **weights** to the canonical weights of the type (if **anisotropic_weights** is empty) or the actual weights with scaling adjustment
-//! \ingroup TasmanianMultiIndexManipulations
-template<typename I>
-void getProperWeights(size_t num_dimensions, I offset, TypeDepth type, const std::vector<I> &anisotropic_weights,
-                      std::vector<I> &weights, I &normalized_offset, bool &known_lower){
-//! Based on the **anisotropic_weights** and selection parameters, create canonical/normalized weights and run basic analysis on the resulting multi-index set.
-//! * **num_dimensions** specifies the number of dimensions of the multi-index
-//! * **type** is the selection type, e.g., type_iptotal
-//! * **anisotropic_weights** is either an empty vector or a vector containing the user specified (non-normalized) weights
-//! * *weights* contains either the normalized weights or the canonical normalized weights
-//! * **normalized_offset** returns the normalized selection threshold
-//! * **known_lower** returns true if the combination of weights and type will guarantee a lower multi-index set
-    if ((type == type_tensor) || (type == type_iptensor) || (type == type_qptensor)){ // special case, full tensor
-        weights.resize(num_dimensions, offset);
-        if (!anisotropic_weights.empty()) for(size_t j=0; j<num_dimensions; j++) weights[j] *= anisotropic_weights[j];
-        normalized_offset = 0; // dummy value, not used
-        known_lower = true; // full-tensors are always lower
-    }else{
-        if (anisotropic_weights.empty()){
-            if ((type == type_curved) || (type == type_ipcurved) || (type == type_qpcurved)){
-                weights.resize(2*num_dimensions);
-                std::fill_n(weights.begin(), num_dimensions, 1);
-                std::fill_n(&(weights[num_dimensions]), num_dimensions, 0);
-            }else{
-                weights.resize(num_dimensions, 1);
-            }
-            normalized_offset = offset;
-            known_lower = true;
-        }else{
-            weights = anisotropic_weights; // copy assign
-            normalized_offset = offset * *std::min_element(weights.begin(), weights.begin() + num_dimensions);
-            if ((type == type_curved) || (type == type_ipcurved) || (type == type_qpcurved)){
-                known_lower = false;
-                for(size_t i=0; i<num_dimensions; i++) if (weights[i] + weights[i+num_dimensions] < 0) known_lower = false;
-            }else{
-                known_lower = true;
-            }
-        }
-    }
-}
+MultiIndexSet selectTensors(size_t num_dimensions, int offset, TypeDepth type, std::function<int(int i)> rule_exactness,
+                            std::vector<int> const &anisotropic_weights, std::vector<int> const &level_limits);
 
 //! \internal
 //! \brief Split a single \b weights vector into two vectors and computes some meta data.
@@ -523,94 +236,6 @@ inline double computeMultiIndexWeight(const std::vector<int> &exactness, const s
         return result;
     }
 }
-
-//! \internal
-//! \brief Generates a multi-index set based on the combination of weights, rule_exactness, offset and type
-//! \ingroup TasmanianMultiIndexManipulations
-template<typename I, typename CacheType, TypeDepth TotalCurvedHyper>
-void generateWeightedTensorsCached(const std::vector<I> &weights, CacheType normalized_offset, std::function<long long(I i)> rule_exactness, MultiIndexSet &mset){
-//! Simplifies the calls to caching and generating multi-indexes
-//! * **I** is the Int type of multi-index (use int)
-//! * **CacheType** should wither match **I** (for level, iptotal and qptotal types) or **double** for curved and hyperbolic rules
-//! * **TotalCurvedHyper** should be either `type_level`, `type_curved` or `type_hyperbolic`, indicating the selection strategy
-//! * **rule_exactness()** handles the cases of quadrature, interpolation or level based selection
-//! * **mset** is the resulting multi-index set
-    size_t num_dimensions = (size_t) mset.getNumDimensions();
-    std::vector<std::vector<CacheType>> cache(num_dimensions);
-    for(size_t j=0; j<num_dimensions; j++){
-        CacheType xi = (CacheType) weights[j]; // anisotropic weight for this direction
-        CacheType eta;
-        if (TotalCurvedHyper == type_curved){
-            eta = (CacheType) weights[j + num_dimensions]; //curved correction
-        }else if (TotalCurvedHyper == type_hyperbolic){
-            eta = 1;
-            if (std::any_of(weights.begin(), weights.end(), [](I w)->bool{ return (w != 1); })) // if not using canonical weights
-                for(auto w : weights) eta += (CacheType) w;
-        }
-        I i = 0;
-        CacheType weight1d = (TotalCurvedHyper == type_hyperbolic) ? 1 : 0;
-        cache[j].push_back(weight1d);
-        do{
-            i++;
-            long long exactness = 1 + rule_exactness(i - 1);
-            if (TotalCurvedHyper == type_level){
-                weight1d = ((CacheType) exactness) * xi;
-            }else if (TotalCurvedHyper == type_curved){
-                weight1d = ((CacheType) exactness) * xi + eta * log1p((CacheType) exactness);
-            }else{
-                weight1d = pow((CacheType) (1 + exactness), xi / eta);
-            }
-            cache[j].push_back(weight1d);
-        }while(ceil(weight1d) <= normalized_offset);
-    }
-    mset = generateLowerMultiIndexSet(num_dimensions,
-                                      [&](const std::vector<int> &index) -> bool{
-                                          CacheType w = 0;
-                                          auto i = index.begin();
-                                          if (TotalCurvedHyper == type_hyperbolic){
-                                              w = 1;
-                                              for(const auto &v : cache) w *= v[*i++];
-                                          }else{
-                                              for(const auto &v : cache) w += v[*i++];
-                                          }
-                                          return (ceil(w) <= normalized_offset);
-                                      });
-}
-
-//! \internal
-//! \brief Generates a multi-index set based on the combination of weights, rule_exactness, offset and type, assume cache has to be build dynamically (non-lower set case)
-//! \ingroup TasmanianMultiIndexManipulations
-template<typename I, typename CacheType>
-void generateWeightedTensorsDynamicCached(const std::vector<I> &weights, CacheType normalized_offset, std::function<long long(I i)> rule_exactness, MultiIndexSet &mset){
-//! Simplifies the calls to caching and generating multi-indexes
-//! * **I** is the Int type of multi-index (use int)
-//! * **CacheType** should wither match **I** (for level, iptotal and qptotal types) or **double** for curved and hyperbolic rules
-//! * **rule_exactness()** handles the cases of quadrature, interpolation or level based selection
-//! * **mset** is the resulting multi-index set
-    size_t num_dimensions = (size_t) mset.getNumDimensions();
-    std::vector<std::vector<CacheType>> cache(num_dimensions);
-    for(size_t j=0; j<num_dimensions; j++) cache[j].push_back(0);
-    mset = generateGeneralMultiIndexSet(num_dimensions,
-                                        [&](const std::vector<I> &index) -> bool{
-                                            CacheType w = 0;
-                                            for(size_t j=0; j<num_dimensions; j++){
-                                                while(index[j] >= (I) cache[j].size()){
-                                                    CacheType xi = (CacheType) weights[j]; // anisotropic weight for this direction
-                                                    CacheType eta = (CacheType) weights[j + num_dimensions];
-                                                    I i = (I) cache[j].size();
-                                                    long long exactness = 1 + rule_exactness(i - 1);
-                                                    cache[j].push_back( ((CacheType) exactness) * xi + eta * log1p((CacheType) exactness) );
-                                                }
-                                                w += cache[j][index[j]];
-                                            }
-                                            return (ceil(w) <= normalized_offset);
-                                        });
-}
-
-//! \internal
-//! \brief Generate the multi-index set defined by the parameters, **rule** cannot be custom
-//! \ingroup TasmanianMultiIndexManipulations
-void selectTensors(int offset, TypeDepth type, std::function<long long(int i)> rule_exactness, const std::vector<int> &anisotropic_weights, MultiIndexSet &set);
 
 //! \internal
 //! \brief Returns a vector that is the sum of entries of each multi-index in the set
